move helper methods to taskhelper
